Channel reworkage, nickname commands
Reworked the way channels are gotten in the bot. For text it's now the first one mentioned in the message. For voice it's a string so voice has to be in quotes unless the name has no spaces.

Commands for renaming multiple people who have bad names all with the same string in it. Removing all nicknames in the guild.