Bug fixes and banned phrase updates
Fixed a bug with being able to disable every single command, bot trying to get the name of a guild it's not longer on, reading preferences, saving banned phrases.

Banned phrases now have punishment types associated with them. You can now have the phrase "discord.gg" be removed and offer no punishment, but have "fuck" be removed and potentially give a role, or kick, or ban.