Added in the last setting
For some reason the text scaling isn't working correctly on the trustedusers textbox and combobox. I don't know why.
Tried to have a dark mode at one point. Did not work well, but keeping the bool in botinfo and its settings commented out in case I ever turn it back on and make it work.